requirements.txt
absl-py==2.0.0
altair==5.1.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.7.2
astroid==2.15.7
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
autopep8==1.5.5
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
Django==4.2.5
dnspython==2.4.2
exceptiongroup==1.1.1
executing==1.2.0
Faker==19.6.2
fastapi==0.95.1
filelock==3.12.4
flake8==3.8.4
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz
gitdb==4.0.10
GitPython==3.1.36
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-datastore==2.18.0
google-cloud-language==2.11.1
google-cloud-logging==3.6.0
google-cloud-spanner==3.40.1
google-cloud-storage==2.11.0
google-cloud-translate==3.12.0
google-cloud-videointelligence==2.11.4
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
h5py==3.9.0
htmltools==0.3.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
install==1.3.5
ipython==8.15.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.17.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.13.1
keyring==24.2.0
keyrings.google-artifactregistry-auth==1.1.2
kiwisolver==1.4.5
lazy-object-proxy==1.6.0
libclang==16.0.6
linkify-it-py==2.0.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
mercurial==5.6.1
more-itertools==10.1.0
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.6.0
pycparser==2.21
pycurl==7.43.0.6
pydantic==1.10.7
pydeck==0.8.0
pydocstyle==6.3.0
pyflakes==2.2.0
Pygments==2.16.1
PyGObject==3.38.0
PyJWT==2.8.0
pylint==2.17.6
pyls-mypy==0.1.8
pymongo==4.5.0
Pympler==1.0.1
PyMySQL==1.1.0
pyodbc==4.0.39
pyparsing==3.1.1
pytest==7.3.1
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-multipart==0.0.6
pytoolconfig==1.2.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rope==1.9.0
rpds-py==0.10.3
rsa==4.9
rsconnect-python @ git+https://github.com/rstudio/rsconnect-python.git@8774dc6dc395f406a63e634e7662245234f50778
SecretStorage==3.3.3
semver==2.13.0
shiny==0.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.46
sqlfaker==0.5
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.26.1
streamlit==1.26.0
supervisor==4.2.2
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.21.1
uvloop==0.17.0
validators==0.22.0
virtualenv==20.24.5
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.2
Werkzeug==2.3.7
wrapt==1.15.0
yapf==0.40.2
zipp==3.17.0